requirements.txt
aiogram==3.4.1
supabase==2.3.4
python-dotenv==1.0.1
httpx>=0.24,<0.26
pydantic>=2.4.1,<2.6
python-dateutil==2.8.2
Pillow==10.2.0
aiofiles==23.2.1 